etl_job/requirements.txt
pymongo
SQLAlchemy<2.0.0
psycopg2-binary
pandas
vaderSentiment